+++ scripts/winesapos-install.sh
@@ -707,8 +707,6 @@ echo "F2FS"
 pacman_install_chroot f2fs-tools
 echo "FAT12, FAT16, and FAT32"
 pacman_install_chroot dosfstools mtools
-echo "FATX16 and FATX32"
-aur_install_chroot fatx
 echo "GFS2"
 aur_install_chroot gfs2-utils
 echo "GlusterFS"