+++ scripts/winesapos-tests.sh
@@ -1167,7 +1167,6 @@ pacman_search_loop \
   erofs-utils \
   exfatprogs \
   f2fs-tools \
-  fatx \
   gfs2-utils \
   glusterfs \
   hfsprogs \